third_party/WebKit/LayoutTests/crypto/wrapKey-lacks-usage-expected.txt

CONSOLE ERROR: key.usages does not permit this operation
Tests that wrapping keys must have the 'wrapKey' usage.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS error is null
PASS successfullyParsed is true

TEST COMPLETE
